Not sure this is going to work.
Eleventy can't do custom template engines yet. They're working on it,
but it's not here yet. Generating the pages as data only goes so far
since I can't put data properties in as front matter... that's really a
template thing. So I guess this is on hold until then? That's
frustrating.