Add the link into the content of the generated link post.
Previously I was showing the link on the rendered template of the page,
but that meant the link didn't show up in the feeds, which seemed dumb.
I *was* going to modify the feeds to include the link if they detected
the `url` property in the post data but that seemed too magical to me.